add more tests to phenotype indexing
